We're a dependency manager, not a package manager
I think “package manager” is way grander in scope, and not really what we're aspiring to. It'd be great if we never have to maintain a centralized database of packages, for example.

Even if we go down that route eventually, I think “dependency management” is a more accurate description of what we're trying to address.