Xcode.swift: Don't unwrap non-optional
Currently, the build fails with the following error:
"Operand of postfix '!' should have optional type; type is 'String'".
The cause is a line in Xcode.swift that attempts to "unwrap"
`fileURL.lastPathComponent`. However, the return type of `-[NSURL
lastPathComponent]` is `String`, not `String?`, so it cannot be
unwrapped.

Remove the unwrapping in order to fix the build.