Stop prematurely disposing of output signals
Because we would dispose of build output subscriptions when the _signal
of signals_ was disposed (not the inner, flattened stream), it would
happen too early, resulting in no build output at all.

Instead, now that RAC is fixed, we can rely upon the lifetime of the
captured sinks we’re writing to. When the sink leaves scope, output
signals (and therefore observations) will automatically disappear too.

Fixes #151.