Fix release downloads by using GitHub API rather than browser URL
Currently, Carthage uses the `browser_url` field on a GitHub release asset to download precompiled binaries. However, when downloading a private binary, this URL only succeeds if the user is logged into their GitHub account on Safari. This because the cookies are shared between Safari and Carthage, and thus the browser URL request is successful because the user is arleady authenticated. However, If the user is not logged into their GitHub account in Safari, the download fails with an opaque error that release zip file is invalid. When you inspect the contents of the downloaded file in the `~/Libarary/Caches/org.carthage.CarthageKit/binaries/` directory, the contents of the file are the string "Not Found", rather than the zipped binary that you'd expect.

To rememdy this, we use the GitHub release asset API with a content type of "application/octet-stream" to download the asset, as described [here](https://developer.github.com/v3/repos/releases/#get-a-single-release-asset). This ensures that the release assets are downloaded successfully regardless of whether the user is logged into Safari.