Build schemes from workspaces, not projects
We were filtering out schemes that don’t actually sit in the given
project. That works great when you’re deciding what to build, but not
when you’re deciding where to build it from.

We should be building from the workspace (if there is one) to correctly
pick up dependencies.

Move the filtering to happen only where we decide which schemes to
build.