Fix precompiled framework platform determination
Currently platform determination of precompiled binaries depends on the architecture of the framework. However, with the addition of watchOS frameworks, this behavior is now incorrect, as both watchOS and iOS frameworks share the `arm` prefix on their architectures, resulting in the behavior documented in #757. To remedy this, instead of relying on architecture, we instead perform the following:

- For each downloaded framework, locate the included Info.plist (location differs in Mac OSX vs. iOS frameworks)
- Within the Info.plist, search for the DTSDKName key
- Strip the version numbers from the SDK name, e.g. `macosx10.10` → `macosx`
- Move the framework to the folder that matches the platform name, as before

Fixes #757

I've tested this with RAC 3.0 and the 4.0 alpha. Let me know if there's any additional cases worth looking into.