Checks is cache folder exits if not it creates it.
We need to check first if the carthage user cache
folder (~/Library/Caches/org.carthage.CarthageKit/)
already exists first before checking if it is writable
by the carthage process.

If it exits we check if it has write permission. If not
we try to create the cache folder with 755 file permissions.