Fix error on `carthage update`
When submodule was already checked out,
`carthage update` stopped with following error:
```
fatal: destination path '…' already exists and is not an empty directory.
```

It was caused in `isGitRepository()`.
The output from `launchGitTask` contained line ending,
so `fileExistsAtPath()` never returned true.
Besides the output pointed relative path of `directoryURL`,
and it was not current directory of `defaultManager()`.