Fix merge order of frameworks to ensure that the device build should be the base (destination) one
The order of SDK list which comes form `xcodebuild -showBuildSettings | grep SUPPORTED_PLATFORMS` are different depending on the platform.

- iOS: "iphonesimulator iphoneos"
- watchOS: "watchsimulator watchos"
- tvOS: "appletvos appletvsimulator"

In tvOS case, the simulator build had been the base framework when merging with the device build. Thus, `CFBundleSupportedPlatforms` would be `AppleTVSimulator` and `DTPlatformName` would be `appletvsimulator` in the resulting Info.plist.

This fix ensures that the device build should be the base one when merging the two frameworks.