Fix dependency resolution when building
As of #1100, dependency resolution for the "build" command uses a comparison sort to resolve ordering of dependencies that need to be built. This change does not build dependencies in the correct order in significantly complex dependency graphs, causing "module not found" errors.

To resolve this, Kahn's topological sort is used in place of a comparison sort to ensure that the dependency graph is built in the correct order.

In our test project with 43 unique dependencies (including transitive dependencies), this change fixed the "module not found" issues. Additionally, since this is a generic algorithm, it can be easily tested (once I get some guidance on a place to put it).