Disable the "Allow debugging when using document Versions Browser" option in the schema
Having this option set would add `-NSDocumentRevisionsDebugMode YES` to the parameters passed to the process when run through XCode (`Cmd-R`). This extra unexpected parameter would cause carthage to fail with the error:
`Unrecognized arguments: -DnouvMtSdbiNmcResg`

Note that the string `-DnouvMtSdbiNmcResg` is an anagram with duplicates letters removed of the string `-NSDocumentRevisionsDebugMode`.