Add --no-clean option to update and build commands
Added to speed up build and update commands allows for incremental builds rather than a full clean and rebuild each time. Default behaviour is as before.

We are using carthage to manage a number of internal components and update frequently and hence the desire to have this option added as time for full rebuilds under Carthage is excessive when regular part of development process.