Ensure that we handle symbolic links correctly when filtering out file paths for projects/workspaces
Carthage’s behavior w/regards to locating projects/workspaces was dependent on the location of the directory the repository was cloned into.

In particular, with macs, if you clone a repository under the /tmp folder, that in fact is a symbolic link to /private/tmp, and will result in incorrect behavior when Carthage attempts to elide paths.

The crux of the problem was the usage of NSFileManager enumeration in locateProjectsInDirectory in conjuction with exclusion paths returned from invoking git commands.
The NSFileManager paths begin with “/private/tmp” while the git commands produce paths that begin with “/tmp”.

The cleanest solution is to use NSFileManager functionality (getRelationship), that tells us correctly if a path is indeed a child of another even in the case of symlinks.

The NSURL extension (hasSubdirectory) does not correctly handle said cases, and it’s really not the appropriate approach in these instances.