Don't copy frameworks produced by dependency projects
A project can have a workspace scheme that builds dependencies located in the
'Carthage/Checkouts' folder.

We want to allow that project to build with the workspace scheme,
but not to copy the resulting frameworks into the Carthage/Build folder,
which would collide with the framework already built by carthage for
that nested dependency.