Remove Xcode version from version files in place of Swift version
Now that we have a way to read Swift versions from frameworks directly, the Xcode version value can now be removed from version files. The Swift version is now read directly from the cached build rather than writing it directly to the file, since Frameworks contain their Swift version.

This will be backwards-compatible with existing version files as they are only losing a key, not gaining a new one.

This was proposed as part of #1850, but it seemed to make sense to split it out into a new PR.