Fix a regression in the new resolver
Still process existing nodes so the edge list gets updated correctly.
Instead, limit the permutation to the previously pinned version of that node.
This way it's still functionally the same as shortcircuiting the node,
but constructs a proper edge set.

Also added a test that failed before this commit in the new resolver