To destination is symlink when archiving on macos
This fixes a workaround for copying frameworks on an APFS disk. When doing an archive build step on macOS, the to destination is a symlink and thus the frameworks were not copied due to the use of COPYFILE_NOFOLLOW. With COPYFILE_NOFOLLOW_SRC the frameworks are copied to the destination.