Ignore anything after a space when parsing semantic versions
According to https://git-scm.com/docs/git-check-ref-format, whitespace are not allowed in a tag, commit SHA, or branch name, so I assume the version ends at the first space.

The reason for this code change is that inline comments in the Cartfile (`#...`) should not be parsed as part of the semantic version. Since comments should at least have a space before the `#` (because otherwise they could be considered part of the commit-ish, `head/foo#2` is a valid tag for example) I decided to use spaces to identify comments. If there is a space and it's not followed by a comment, the file parser will throw an error because it expects either a comment or the line to be over.