Refactor detection of Cartfile comments
## Problem
Having the detection of Cartfile comments inside the version parser makes the parser unnecessarily tied to the Cartfile syntax. Additionally, since there is no case in which we would not parse a version speficier and not ignore the comments, either we have to live with a parameter that is always set to `true` in the entire codebase, or remove the parameter and make strong assumptions on how the version parsing is used.

## Solution
This PR moves comment filtering to the Cartfile parsing code. Comments are stripped from each line before passing lines to the version parser.

The detection of comments is done with the following strategy:
- any `#` character that is not inside "quotes" is considered the beginning of the comment
- for the purpose of determining whether a character is inside quotes, the code relies on the fact that there should not be nested quotes (e.g. `quotes "like \"this\""`) inside a cartfile specification.

Even if not necessary at the moment, as the comment indicator is a single character, the implementation is designed to handle multi-character comment indicators. This has influenced the choice of using `String.components(separatedBy:)` over other solutions.

All the above is documented in code comments, and the PR includes tests to cover the comment stripping.