Pad `install_name_tool`-modified headers to appease Homebrew.
Homebrew (as of version 1.7.6¹) passes those as `LDFLAGS` to enable their use of `install_name_tool`², but, for our use of Swift via xcodebuild, `OTHER_LDFLAGS` is needed — so do this, at least temporarily.

The placement of `OTHER_LDFLAGS` in `XCODEFLAGS` can change after the formula for Carthage 0.31.1 is published to Homebrew. But, for now, it stays as it in order to enable `brew install --head carthage` to not fail on the [`inreplace`](https://github.com/Homebrew/homebrew-core/blob/85ea337/Formula/carthage.rb#L19-L22) which matches everything in that line preceding the space before `OTHER_LDFLAGS`. The formula for Carthage 0.31.1 will remove that `inreplace`.

⑴ See <https://github.com/Homebrew/brew/releases/tag/1.7.6>.

⑵ See <https://github.com/Homebrew/legacy-homebrew/issues/20233>.