Set an invalid swift version in ProjectSpec to force the dSYM to be checked
Previously, this example for "createAndCheckVersionFiles" relied on the
underlying preferred encoding of Swift strings. The test uses
`NSString.fastestEncoding` to rewrite a header, but Carthage assumes
utf8 encoding when it reads the header in `Xcode.frameworkSwiftVersion(_:)`.

Prior to Swift 5, Strings preferred utf16. This meant that the test
would overwrite the header in utf16, and Carthage would fail to parse
it, leading to the fallback behavior behing verified by the test. Now
that Strings are backed by utf8 by default[1], the overwritten
header would still be parsed by Carthage and interpreted as a valid
older Swift version.

The fix is to change the version string in the header to "<invalid>".
This causes Carthage to fail to parse the rewritten header by violating
the expected format of the version string.

[1]: https://swift.org/blog/utf8-string/