Prevent directory traversal (in more piecemeal way) allowing «project-directory» flag to function.
Also, sanitize [nul characters](https://en.wikipedia.org/wiki/Null_character) in `GitURL`s (transposing them to «␀».)

The cases in the modified test are — given a URL is specified and provided is a string without URL scheme — pathological, and as such the result for them from `GitURL.name` are maybe unintuitive, but decidedly safe.

The flag «--project-directory» clashed with the formerly placed (but never shipped in a tagged Carthage release) invocation of `FileManager.currentDirectoryPath`.

To avoid the clash, we settle on a less intuitive, but most importantly safe change — _to be clear_ a *breaking change* for some pathological URL-scheme–less or nul-character–containing Cartfile lines and `GitURL`s.