Fix invalid lipo command in stripArchitectures() when embedding dSYMs (#3155)
If the architectures argument is empty (because there are no architectures to be stripped), this method will produce and execute an invalid lipo command because it will contain no -remove arguments. To fix this, I make the method a no-op if the input list of architectures is empty.

This only impacts `carthage copy-frameworks` when embedding dSYMs. We fixed a similar issue for framework embedding in 58551c05f40703512a8f53d8a169b21b6774164b.